Added last_closer to Inline object.
This helps us avoid unneeded backtracking in pathological input
of the form:

*a
**a
*a
**a
*a
etc.

If we get to position k without finding a closing delimiter,
then backtrack to 1, we can assume we won't find a closing
delimiter when parsing forward again.

This could no doubt be polished up, e.g. by making it sensitive
to the kind of delimiter.